Added imgur gallery functionality and markwon link handling
