+++ app/src/main/java/me/cniekirk/flex/data/remote/ImgurApi.kt
@@ -12,13 +12,13 @@ interface ImgurApi {
     @POST("3/upload")
     suspend fun uploadImage(
         @Part image: MultipartBody.Part? = null,
-        @Header("Client-ID") clientId: String = "a3a83e144ed48df"
+        @Header("Authorization") authorization: String = "Client-ID a3a83e144ed48df"
     ): ImgurResponse<Data>
 
     @GET("3/album/{albumHash}/images")
     suspend fun getGalleryImages(
         @Path("albumHash") albumHash: String,
-        @Header("Client-ID") clientId: String = "a3a83e144ed48df"
+        @Header("Authorization") authorization: String = "Client-ID a3a83e144ed48df"
     ): ImgurResponse<List<ImageData>>
 
 }
