+++ app/src/main/java/me/cniekirk/flex/di/DataStoreModule.kt
@@ -8,7 +8,9 @@ import androidx.datastore.dataStoreFile
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.android.scopes.ViewModelScoped
 import dagger.hilt.components.SingletonComponent
 import me.cniekirk.flex.FlexSettings
 import me.cniekirk.flex.data.local.prefs.SettingsSerializer
